feat: finish authentication logic
