+++ src/actions/authActions/index.ts
@@ -1,20 +1,21 @@
 import { mapAuthResponseToUser } from "_/helpers"
 import { addUser } from "_/store/authStore"
-import { AuthCredentials, Position, User } from "_/types"
+import { AuthCredentials } from "_/types"
 import { AppThunk } from "_/types/appThunk"
 
-export const authenticateAction = (credentials: AuthCredentials, position: Position): AppThunk => {
+export const authenticateAction = (credentials: AuthCredentials): AppThunk => {
 
-    return async (dispatch, getState, { authService, userDatabaseRepository }) => {
+    return async (dispatch, getState, { authService, userService }) => {
 
         const authResponse = await authService.authenticateGithub(credentials)
         if (!authResponse) throw new Error("Erro na autenticação com Git.")
+
+        const position = await userService.getUserPosition()
         const newUser = mapAuthResponseToUser(authResponse, position)
 
         dispatch(addUser(newUser))
+        await userService.createUser(newUser)
 
-        const fUser = mapUserToFirebaseUser(newUser)
-        await userDatabaseRepository.createOrReplace(fUser);
     }
 }
 
@@ -22,22 +23,6 @@ export const authenticateAction = (credentials: AuthCredentials, position: Posit
 
 
 
-export interface FirebaseUserDto {
-    username: string
-    id: number
-    photoUrl?: string
-    email: string
-    techs?: Array<string>
-}
 
-export const mapUserToFirebaseUser = (user: User): FirebaseUserDto => {
-    return {
-        email: user.email || "",
-        id: user.id,
-        username: user.username,
-        photoUrl: user.photoUrl,
-        techs: user.techs
-    }
-}
