+++ src/adapters/index.ts
@@ -1,1 +1,2 @@
 export * from "./http"
+export * from "./location"
