+++ src/adapters/location/index.ts
@@ -1,0 +1,2 @@
+export * from "./location"
+export * from "./types"
