+++ src/services/authService/types.ts
@@ -1,0 +1,5 @@
+import { AuthCredentials, AuthResponse } from "_/types";
+
+export interface AuthService {
+    authenticateGithub: (credentials: AuthCredentials) => Promise<AuthResponse | undefined>
+}
