+++ src/services/userService/dto/firebaseUserDto.ts
@@ -1,0 +1,13 @@
+export interface FirebaseUserDto {
+    username: string
+    id: number
+    photoUrl?: string
+    email?: string
+    techs?: Array<string>
+    profileUrl: string
+    position: {
+        latitude: number
+        longitude: number
+        geohash: string
+    }
+}
