+++ src/services/userService/types.ts
@@ -1,0 +1,6 @@
+import { Position, User } from "_/types";
+
+export interface UserService {
+    getUserPosition: () => Promise<Position>
+    createUser: (user: User) => Promise<void>
+}
